<Update>:<아이템 정보 팝업 개선. 각종 스크립트에 로컬라이즈 적용>
<description>
 1. 변경된 Ability와 Effect Variables에 맞추어 ItemInfoPopup 개선.
 2. 추가된 Artifact Item, Battle Scene 테스트 진행중.
 3. Buff Type Artifact 테스트 진행 완료. 로직 정상작동 확인.
 4. Debuff Type과 Trigger Type Artifact skill의 위치지정 기능 스크립트 작성.
 5. Artifact Effect Active Slot 로직 개선과 복잡했던 Init로직 정리.

<issue-number>
# 연관된 이슈 첨부, 여러 개 추가 가능

+++ ArrowDefence_Project/Assets/02.Scripts/CodingCat_Script/00.Battle Script/AccessoryEffect/ArtifactCondition.cs
@@ -23,7 +23,7 @@
                 return conditionType;
             }
         }
-        public float MaxStack {
+        public int MaxStack {
             get => maxStack;
         }
         public float MaxCost {
